# swagger2markup/src/test/resources/json/swagger.json
{"swagger":"2.0","info":{"description":"Speakers API Description","version":"1.0.0","title":"Swagger Speakers","contact":{"name":"Tsypuk Roman","url":"https://tsypuk.github.io/springrestdoc","email":"tsypuk.rb@gmail.com"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"topic-controller","description":"Topic Controller"},{"name":"Speakers","description":"Operations about speaker"}],"paths":{"/speakers":{"get":{"tags":["Speakers"],"summary":"allSpeakers","operationId":"allSpeakersUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Resources«Speaker Resource»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/speakers/{id}":{"get":{"tags":["Speakers"],"summary":"Find speaker by ID","description":"For valid response try integer IDs with value 1 ... 999. Other values will generated exceptions","operationId":"getSpeakerUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"ID of speaker that needs to be fetched","required":true,"type":"integer","maximum":999.0,"minimum":1.0,"format":"int64"}],"responses":{"200":{"description":"Successful retrieve the speaker.","schema":{"$ref":"#/definitions/Speaker Resource"}},"400":{"description":"Invalid ID supplied"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error."},"403":{"description":"Forbidden"},"404":{"description":"Speaker not found"}}},"post":{"tags":["Speakers"],"summary":"createSpeaker","operationId":"createSpeakerUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Speaker Resource"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["Speakers"],"summary":"updateSpeaker","operationId":"updateSpeakerUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["Speakers"],"summary":"deleteSpeaker","operationId":"deleteSpeakerUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}}},"/topics/{id}":{"get":{"tags":["topic-controller"],"summary":"getTopic","operationId":"getTopicUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TopicResource"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"Resources«Speaker Resource»":{"type":"object","properties":{"_embedded":{"$ref":"#/definitions/Collection«Speaker Resource»"},"_links":{"type":"array","items":{"$ref":"#/definitions/Link"}}}},"TopicResource":{"type":"object","properties":{"_links":{"type":"array","items":{"$ref":"#/definitions/Link"}},"description":{"type":"string"},"duration":{"$ref":"#/definitions/Duration"},"name":{"type":"string"}}},"Collection«Speaker Resource»":{"type":"object"},"TemporalUnit":{"type":"object","properties":{"dateBased":{"type":"boolean"},"duration":{"$ref":"#/definitions/Duration"},"durationEstimated":{"type":"boolean"},"timeBased":{"type":"boolean"}}},"Duration":{"type":"object","properties":{"nano":{"type":"integer","format":"int32"},"negative":{"type":"boolean"},"seconds":{"type":"integer","format":"int64"},"units":{"type":"array","items":{"$ref":"#/definitions/TemporalUnit"}},"zero":{"type":"boolean"}}},"Speaker Resource":{"type":"object","required":["company","name","status"],"properties":{"_links":{"type":"array","items":{"$ref":"#/definitions/Link"}},"company":{"type":"string","example":"Lohika","description":"The company that speaker is working on.","enum":["Full company names."]},"name":{"type":"string","example":"Roman Tsypuk","description":"Name of the Speaker in full format."},"status":{"type":"string","example":"I like Spring & Rest Docs.","description":"Current status.","enum":["Only positive :)"]}},"description":"Speaker detailed description."},"Link":{"type":"object","properties":{"href":{"type":"string"},"rel":{"type":"string"},"templated":{"type":"boolean"}}}}}